Update README.md
🎮
Rock-paper-scissors is a simple hand game usually played between two people. Each player simultaneously forms one of three shapes with an outstretched hand. The possible shapes are "rock" (a fist), "paper" (an open hand), and "scissors" (a fist with the index and middle fingers extended, forming a V). The winner is determined by the rules: rock beats scissors, scissors beats paper, and paper beats rock. It's a game of chance often used as a decision-making tool or a fun activity. ENJOY THE GAME 🎮